+++ test_app/tests/authentication/management/test_authenticators.py
@@ -97,7 +97,6 @@ def test_authenticators_cli_initialize(
     err = StringIO()
 
     # Sanity check:
-    # _system user exists
     assert django_user_model.objects.count() == 0
 
     # Optionally create admin user
